+++ source/helpers/jquery.fancybox-buttons.js
@@ -1,6 +1,6 @@
  /*!
  * Buttons helper for fancyBox
- * version: 1.0.3
+ * version: 1.0.4 (Mon, 01 Oct 2012)
  * @requires fancyBox v2.0 or later
  *
  * Usage:
@@ -12,10 +12,6 @@
  *         }
  *     });
  *
- * Options:
- *     tpl - HTML template
- *     position - 'top' or 'bottom'
- *
  */
 (function ($) {
 	//Shortcut for fancyBox object
@@ -23,7 +19,11 @@
 
 	//Add helper object
 	F.helpers.buttons = {
-		tpl  : '<div id="fancybox-buttons"><ul><li><a class="btnPrev" title="Previous" href="javascript:;"></a></li><li><a class="btnPlay" title="Start slideshow" href="javascript:;"></a></li><li><a class="btnNext" title="Next" href="javascript:;"></a></li><li><a class="btnToggle" title="Toggle size" href="javascript:;"></a></li><li><a class="btnPrint" title="Print" href="javascript:;">print</a></li><li><a class="btnClose" title="Close" href="javascript:jQuery.fancybox.close();"></a></li></ul></div>',
+		defaults : {
+			skipSingle : false, // disables if gallery contains single image
+			position   : 'top', // 'top' or 'bottom'
+			tpl        : '<div id="fancybox-buttons"><ul><li><a class="btnPrev" title="Previous" href="javascript:;"></a></li><li><a class="btnPlay" title="Start slideshow" href="javascript:;"></a></li><li><a class="btnNext" title="Next" href="javascript:;"></a></li><li><a class="btnToggle" title="Toggle size" href="javascript:;"></a></li><li><a class="btnPrint" title="Print" href="javascript:;">print</a></li><li><a class="btnClose" title="Close" href="javascript:jQuery.fancybox.close();"></a></li></ul></div>',
+		},
 		list : null,
 		buttons: null,
 
@@ -57,7 +57,7 @@
 			var buttons = this.buttons;
 
 			if (!buttons) {
-				this.list = $(opts.tpl || this.tpl).addClass(opts.position || 'top').appendTo('body');
+				this.list = $(opts.tpl).addClass(opts.position).appendTo('body');
 
 				buttons = {
 					prev   : this.list.find('.btnPrev').click( F.prev ),